lsdb-dataframes.ipynb: save dask warnings/errors w. large loads
It looks like Dask communicates in the same thread it does work, leading to
timeouts with large workloads. It also looks like the cluster recovers from
this, but it throws many, many, scary errors/warnings/exceptions along the
way messages. They're saved in the notebook in this commit, for alter
inspection.