ephem-service/astcheck: functional REST-ful API prototype
astcheck.py is the main module and doubles as a command line utility.
service.py is the FastAPI service module.

To start the service, run:

	./astcheck.py serve <cache_file.pkl>

(this will bind to localhost:8000 by default).

To query the service, run:

	./astcheck.py query 60853.1 32 11 --radius=1.8

(this will query localhost:8000 by default).

Data exchange is done via Arrow-serialied tensors/recordbatches (for speed).